Slightly improve pagination logic
Prior to this commit we'd naively link to a page that we could have
known didn't exist. This commit check for the existence of the folder
and only paginates to pages we know exist.

The problem, of course, is the year boundaries. The last page of one
year can't know about the first page of the next year without changing
around how this code works. For now, I don't care, but it's something
that I should fix.