Add basic rackup config
With this commit you can serve the contents of the build folder like
this:

```ruby
$ rackup
```

It'll use the `config.ru` file in the root of the project, but consider
the `build` folder the root of the site it's serving. The very general
`Rack::Static` config means that all content is matched and served from
that location.