WIP
This is my first commit after coming back to this branch months later.
What I found was that the Resolver wasn't fleshed out and there was a
mistaken on the Dropbox path we were listening to. I also updated the
ignore file so that this stuff isn't showing up in the diff anymore.